Move PSClassDeclaration to its own package
