+++ src/main/java/org/purescript/psi/classes/PSClassDeclaration.kt
@@ -1,5 +1,6 @@
-package org.purescript.psi
+package org.purescript.psi.classes
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.PSPsiElement
 
 class PSClassDeclaration(node: ASTNode) : PSPsiElement(node)
